add Envelope class

+++ include/trackerboy/synth/NoiseGen.hpp
@@ -28,7 +28,7 @@ private:
     unsigned mShiftCounter;
     unsigned mShiftCounterMax;
 
-    float mStepsPerSample;
+    unsigned mStepsPerSample;
 };
